infosecwriteups/requirement.txt
python-dotenv
requests
streamlit
pandas
plotly
